requirements.txt
Flask==1.1.2
google-cloud-datastore==1.12.0
requests==2.22.0
json2html==1.3.0
google_auth==1.14.3
requests_oauthlib==1.3.0